feat: コードの共通化、不要コードの削除
