+++ src/tufspot/resources/views/livewire/paginated-post-list.blade.php
